requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.0
celery==5.4.0
certifi==2024.8.30
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
cryptography==43.0.0
dj-database-url==2.2.0
Django==5.1
django-celery-beat==2.7.0
django-extensions==3.2.3
django-heroku==0.3.1
django-ratelimit==4.1.0
django-sslserver==0.22
django-timezone-field==7.0
django-widget-tweaks==1.5.0
django_csp==3.8
gunicorn==23.0.0
idna==3.8
kombu==5.4.2
packaging==24.1
prompt_toolkit==3.0.47
psycopg2==2.9.9
psycopg2-binary==2.9.9
pycparser==2.22
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
stripe==10.11.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.7.0